backend/requirements.txt
feedparser==6.0.11
supabase==2.3.4
python-dotenv==1.0.0
google-genai
fastapi==0.115.0
uvicorn[standard]==0.32.0
